NIFI-9953 - Updated license files, fixed checkstyle issues.
