+++ nifi-toolkit/nifi-property-encryptor-tool/src/main/java/org/apache/nifi/util/console/PropertyEncryptorCLI.java
@@ -16,6 +16,8 @@
  */
 package org.apache.nifi.util.console;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import picocli.CommandLine;
 
 @CommandLine.Command(name = "PropertyEncryptor",
@@ -28,41 +30,14 @@ import picocli.CommandLine;
 )
 class PropertyEncryptorCLI extends DefaultCLIOptions implements Runnable {
 
-    /**
-     * ./property-encryptor.sh --help
-     * usage: org.apache.nifi.toolkit.propertyencryptor.PropertyEncryptorMain [-h] [-v] [encrypt | decrypt | migrate | translate-cli]
-     *
-     * This tool can be used to easily encrypt configuration files for NiFi and its sub-projects (NiFi Registry, MiNiFi), as well as the flow.xml.gz or flow.json.gz files. Given a root directory, password and scheme it will protect all secret values within configuration files or within the flow.xml.gz/flow.json.gz with the key/password. The tool can also decrypt configuration files given the correct credentials. It also allows for migrating the password used from old to new, and changing the encryption scheme used.
-     *
-     * -h,--help           Show usage information (this message)
-     * -v,--verbose        Sets verbose mode (default false)
-     *
-     * Command examples:
-     * ./property-encryptor.sh encrypt [config | flow] [root-nifi-dir | root-nifi-registry-dir | root-minifi-dir] [password] [scheme]
-     * ./property-encryptor.sh decrypt config [root-nifi-dir | root-nifi-registry-dir | root-minifi-dir] [password] [scheme]
-     * ./property-encryptor.sh migrate [config | flow] [root-nifi-dir | root-nifi-registry-dir | root-minifi-dir] [new-password] [new-scheme]
-     * ./property-encryptor.sh translate-for-nifi-cli nifi.properties
-     *
-     * Encrypt examples:
-     * ./property-encryptor.sh encrypt config root-nifi-dir passphrase AES-GCM
-     *
-     * OR
-     *
-     * ./property-encryptor.sh config <encrypt | decrypt | migrate>
-     * ./property-encryptor.sh flow <encrypt | migrate> base-nifi-dir password [PBECipherProvider] [EncryptionMethod]
-     * ./property-encryptor.sh translate-for-nifi-cli nifi.properties
-     *
-     *                                  sub-subcommand
-     * ./property-encryptor.sh encrypt [config | flow]
-     *
-     */
+    private static final Logger logger = LoggerFactory.getLogger(PropertyEncryptorCLI.class);
 
     @Override
     public void run() {
         if (verboseLogging) {
-            System.out.println("Verbose logging enabled");
+            logger.info("Verbose logging enabled");
         }
-        System.out.println("Running the Property Encryptor...");
+        logger.info("Running the Property Encryptor..");
     }
 
     public static void main(String[] args) {
