+++ nifi-toolkit/nifi-property-encryptor-tool/src/main/java/org/apache/nifi/util/file/ConfigurationFileResolver.java
@@ -1,3 +1,19 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package org.apache.nifi.util.file;
 
 import org.apache.nifi.properties.ApplicationProperties;
@@ -23,7 +39,7 @@ public class ConfigurationFileResolver {
     /**
      * Use the nifi.properties file to locate configuration files referenced by properties in the file
      *
-     * @return
+     * @return List of application configuration files
      */
     public List<File> resolveConfigurationFilesFromApplicationProperties(final ApplicationProperties applicationProperties) throws ConfigurationFileResolverException {
         ArrayList<File> configurationFiles = new ArrayList<>();
@@ -64,8 +80,6 @@ public class ConfigurationFileResolver {
 
     /**
      * Return a configuration file absolute path based on the confDirectory rather than Java's working path
-     * @param relativeFile
-     * @return
      */
     private File getAbsolute(final File relativeFile) {
         if (relativeFile.isAbsolute() ) {
