+++ nifi-toolkit/nifi-property-encryptor-tool/src/main/java/org/apache/nifi/util/file/FileUtilities.java
@@ -1,3 +1,19 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package org.apache.nifi.util.file;
 
 import java.io.File;
@@ -47,14 +63,15 @@ public class FileUtilities {
         } else if (directoryContainsFilename(baseDirectory, DEFAULT_CONF_DIR)) {
             return getAbsolutePath(getDefaultConfDirectory(baseDirectory).toPath());
         } else {
-            throw new IllegalArgumentException(String.format("The configuration directory [%s]/ could not be found within [%s] or it did not contain a properties file", DEFAULT_CONF_DIR, baseDirectory));
+            throw new IllegalArgumentException(
+                    String.format("The configuration directory [%s]/ could not be found within [%s] or it did not contain a properties file", DEFAULT_CONF_DIR, baseDirectory));
         }
     }
 
     /**
      * Get the properties file either NiFi or NiFi Registry from within the configuration directory
-     * @param confDirectory
-     * @return
+     * @param confDirectory The ./conf directory
+     * @return The NiFi or NiFi Registry properties file (eg. nifi.properties or nifi-registry.properties)
      */
     public static File resolvePropertiesFile(final Path confDirectory) {
         if (directoryContainsFilename(confDirectory, NIFI_PROPERTIES_DEFAULT_NAME)) {
