--- nifi-toolkit/nifi-property-encryptor-tool/src/main/java/org/apache/nifi/xml/XMLCryptoException.java
@@ -1,11 +1,0 @@
-package org.apache.nifi.xml;
-
-public class XMLCryptoException extends RuntimeException {
-    public XMLCryptoException(final String message) {
-        super(message);
-    }
-
-    public XMLCryptoException(final String message, final Throwable cause) {
-        super(message, cause);
-    }
-}
